fix connection failure handling
clientConnectionFailed() isn't appropriate here, since the connection
actually did succeed. it's supposed to take an argument implementing
IConnector, too, so passing the ManagedThriftClientProtocol instance
mucks things up later if a connection retry is attempted.

add a call, clientSetupFailed(), to handle this situation, and have
the client's transport loseConnection() so that the factory's clientGone
will be called.