improve on_insufficient_conns handling
have all pending requests in the queue when on_insufficient_conns is
called, including the most recently-submitted. it is no longer
appropriate to throw an exception from on_insufficient_conns to cancel
the most recently-submitted request, but it can instead be cancelled
from within the request queue.

on_insufficient_conns will now also be called when there are pending
requests and the last valid connection is lost.