fix thrift's handling of connection loss
the thrift-generated version of connectionLost hands an error to a
series of callbacks (which are likely to invoke other callbacks) but
does not account for the possibility that its own request registry could
change in the course of handling the error thrown.

this version fixes that by popping elements one at a time from the
registry, without trying to use a single iterator.

Avoids the 'dictionary changed size during iteration' errors.