fix misordered request failure/factory stop
when a connection failed at the right place, and the protocol object's
connectionLost method was called before the factory's
stop_working_on_queue(), and the protocol had a pending request to
errback, the processing chain was causing the factory to try to fetch
a new request from the queue before it knew that its own connection was
kaput. This caused hijinks in the queue handling.

now put off the determination of whether to ask for more work from the
service, until after all pending callbacks complete.